Remove duplicate api_helpers file
